AmazonFireTV_Smart_LED_TV_2K-7.6.9.0-20250314 [bootable]

+++ bootable/bootloader/uboot-amlogic/t950x4/bl33/common/main.c
@@ -271,7 +271,7 @@ void main_loop(void)
 		if (!is_transition_done) {
 			ret = 0;
 			idme_get_var_external("oem_data", oem_data, sizeof(oem_data));
-			if (strstr(oem_data, "hadrian") != NULL || strstr(oem_data, "shine-tm") != NULL) {
+			if (strstr(oem_data, "hadrian") != NULL || strstr(oem_data, "shine-tm") != NULL || strstr(oem_data, "shine-jb") != NULL || strstr(oem_data, "shine-jm") != NULL || strstr(oem_data, "almondmo") != NULL || strstr(oem_data, "almondh") != NULL || strstr(oem_data, "almondb") != NULL) {
 				ret += run_command("amlmmc erase dfs", 0);
 				ret += run_command("amlmmc erase dkernel", 0);
 				ret += run_command("amlmmc erase diag_userdata", 0);
@@ -302,7 +302,7 @@ void main_loop(void)
 				printf("Transition from Diag to FOS failed\n");
 			} else {
 				printf("Transition from Diag to FOS succeed\n");
-				if (strstr(oem_data, "almond") != NULL || strstr(oem_data, "shine-jt") != NULL || strstr(oem_data, "dahlia") != NULL) {
+				if ((strstr(oem_data, "almond") != NULL && strstr(oem_data, "almondmo") == NULL && strstr(oem_data, "almondh") == NULL && strstr(oem_data, "almondb") == NULL) || strstr(oem_data, "shine-jt") != NULL || strstr(oem_data, "dahlia") != NULL) {
 					if (bootmode == IDME_BOOTMODE_TRANSITION) {
 						run_command("reboot", 0);
 					} else {
@@ -322,7 +322,7 @@ void main_loop(void)
 					while (1)
 						udelay(1000*1000);
 					}
-				} else if (strstr(oem_data, "hadrian") != NULL || strstr(oem_data, "shine-tm")) {
+				} else if (strstr(oem_data, "hadrian") != NULL || strstr(oem_data, "shine-tm") || strstr(oem_data, "shine-jb") || strstr(oem_data, "shine-jm") || strstr(oem_data, "almondmo") || strstr(oem_data, "almondh") || strstr(oem_data, "almondb")) {
 					ret = 0;
 					ret += run_command("osd open", 0);
 					ret += run_command("osd clear", 0);
